Began migrating release generation code to a node based program
The reason for migration is to have less load on the server itself. Although the process is not too costly for the server to handle, executing movie generation code on the server with php has become inconsistent and unreliable with issues such as timeouts. Since the code will not be in production, I have opted to make the generation it's own small program.
In this new programs current state, it uses some placeholder values for values such as the following: tagline, description, watch_link, languages and categories

Login values for the database do not belong on the Git repo for security reasons, documentation will be found in the read me file at a later date